feat: 调整增量式PI控制器输出类型为浮点，更新Alpha-Beta滤波器参数以优化控制响应

+++ BALANCE/Inc/balance_task.h
@@ -64,7 +64,7 @@ typedef struct{
 	float Bias;    // 与目标值偏差
 	float LastBias;// 上一次的偏差
 	float LastestBias;//上两次的偏差
-	int Output;    // 输出
+	float Output;  // 输出(内部使用浮点累计,避免小量被截断)
 	float kp;      // kp值
 	float ki;      // ki值
 	float kd;      // kd值
